First commit to re-organize the tests by protocol
