+++ tests/ethercat/test_ethercat_dictionary.py
@@ -1,6 +1,9 @@
+
+import pytest
+
 from ingenialink.dictionary_deprecated import Dictionary
 
-
+@pytest.mark.ethercat
 def test_dictionary(connect_to_slave):
     servo, net = connect_to_slave
     assert servo is not None and net is not None
