+++ tests/ethercat/test_ethercat_network.py
@@ -5,6 +5,15 @@ from ingenialink.exceptions import ILFirmwareLoadError
 
 
 @pytest.mark.ethercat
+def test_connect_to_slave(connect_to_slave):
+    servo, net = connect_to_slave
+    assert servo is not None and net is not None
+    assert len(net.servos) == 1
+    fw_version = servo.read('DRV_ID_SOFTWARE_VERSION')
+    assert fw_version is not None and fw_version != ''
+
+
+@pytest.mark.ethercat
 def test_scan_slaves(read_config):
     net = EthercatNetwork(read_config['ethercat']['ifname'])
